backend/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
blinker>=1.6.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.5
MarkupSafe==2.1.3
packaging==24.1
#psycopg2-binary==2.9.9
psycopg2-binary>=2.9,<3

PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.0
flask_mail